Renamed matchers to use a Matcher suffix.
