+++ src/matchers/IpMatcher.js
@@ -4,7 +4,7 @@
  * @flow
  */
 
-import UrlMatcher from './Url';
+import UrlMatcher from './UrlMatcher';
 import { IP_PATTERN } from '../constants';
 
 import type { MatchResponse, MatcherFactory } from '../types';
@@ -12,7 +12,7 @@ import type { MatchResponse, MatcherFactory } from '../types';
 const IP_REGEX = new RegExp(IP_PATTERN, 'i');
 
 export default class IpMatcher extends UrlMatcher {
-  constructor(name: string, options: Object = {}, factory: ?MatcherFactory = null) {
+  constructor(name: string, options?: Object = {}, factory?: ?MatcherFactory = null) {
     super(name, {
       ...options,
       // IPs dont have TLDs
