+++ src/matchers/UrlMatcher.js
@@ -16,7 +16,7 @@ const URL_REGEX = new RegExp(URL_PATTERN, 'i');
 export default class UrlMatcher extends Matcher<UrlOptions> {
   options: UrlOptions;
 
-  constructor(name: string, options: Object = {}, factory: ?MatcherFactory = null) {
+  constructor(name: string, options?: Object = {}, factory?: ?MatcherFactory = null) {
     super(name, {
       customTLDs: [],
       validateTLD: true,
@@ -24,7 +24,7 @@ export default class UrlMatcher extends Matcher<UrlOptions> {
     }, factory);
   }
 
-  replaceWith(match: string, props: Object = {}): ReactNode<*> {
+  replaceWith(match: string, props?: Object = {}): ReactNode<*> {
     if (this.options.validateTLD) {
       const { host } = props.urlParts;
       const validList = TOP_LEVEL_TLDS.concat(this.options.customTLDs);
