+++ tests/matchers/UrlMatcher.test.js
@@ -1,5 +1,5 @@
 import Parser from '../../src/Parser';
-import UrlMatcher from '../../src/matchers/Url';
+import UrlMatcher from '../../src/matchers/UrlMatcher';
 import { URL_PATTERN } from '../../src/constants';
 import { TOKEN_LOCATIONS, createExpectedTokenLocations, parentConfig } from '../mocks';
 
@@ -96,7 +96,7 @@ const INVALID_URLS = [
   { url: 'sftp://domain.com' },
 ];
 
-describe('matchers/Url', () => {
+describe('matchers/UrlMatcher', () => {
   let matcher = new UrlMatcher('url');
   const pattern = new RegExp(`^${URL_PATTERN}$`, 'i');
