+++ tests/mocks.js
@@ -55,7 +55,7 @@ export function createExpectedTokenLocations(value, factory, flatten = false) {
 export const VALID_EMOJIS = [
   ['1F468-200D-1F469-200D-1F467-200D-1F466', '👨‍👩‍👧‍👦', ':family_mwgb:'],
   ['1F1FA-1F1F8', '🇺🇸', ':flag_us:'],
-  ['1F621', '😡', ':rage:'],
+  ['1F621', '😡', ':enraged:'],
   ['1F63A', '😺', ':smiling_cat:'],
   ['1F3EF', '🏯', ':japanese_castle:'],
   ['1F681', '🚁', ':helicopter:'],
